ログアウトボタン追加 snsのログアウトボタン削除
