+++ src/app/damage-calculator/page.js
@@ -338,7 +338,7 @@ const DamageCalculatorPage = () => {
                     zIndex: 1000,
                 }}
             >
-                <Link href="/">
+                <Link href="/top">
                     <Image
                         src="/images/gaming.gif"
                         width={50}
