+++ src/app/page.js
@@ -1,49 +1,115 @@
 "use client";
 
+import React, { useState, useEffect } from "react";
 import Link from "next/link";
 import Image from "next/image";
+import { getAuth, signInWithPopup, GoogleAuthProvider, signInWithEmailAndPassword, onAuthStateChanged, signOut } from "firebase/auth";
+import { initializeApp } from "firebase/app";
+import Cookies from "js-cookie";
 
-const TopPage = () => {
+// Firebaseの設定情報
+const firebaseConfig = {
+    apiKey: process.env.NEXT_PUBLIC_API_KEY,
+    authDomain: process.env.NEXT_PUBLIC_AUTH,
+    projectId: process.env.NEXT_PUBLIC_PRO,
+    storageBucket: process.env.NEXT_PUBLIC_BUKET,
+    messagingSenderId: process.env.NEXT_PUBLIC_POKEMON_ID,
+    appId: process.env.NEXT_PUBLIC_APP_ID,
+    measurementId: process.env.NEXT_PUBLIC_MEASUREMENT_ID,
+};
+
+// Firebaseアプリの初期化
+const app = initializeApp(firebaseConfig);
+const auth = getAuth(app);
+const provider = new GoogleAuthProvider();
+
+export default function LoginPage() {
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const [error, setError] = useState("");
+    const [user, setUser] = useState(null);
+
+    // Firebaseの認証状態を監視
+    useEffect(() => {
+        const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
+            if (currentUser) {
+                setUser(currentUser);
+                Cookies.set("user", JSON.stringify(currentUser), { expires: 7 });
+                window.location.href = "/top";
+            } else {
+                setUser(null);
+            }
+        });
+
+        return () => unsubscribe();
+    }, []);
+
+    // メールアドレスとパスワードでログイン
+    const handleLogin = async (e) => {
+        e.preventDefault();
+        setError("");
+
+        if (!email || !password) {
+            setError("メールアドレスとパスワードを入力してください。");
+            return;
+        }
+
+        try {
+            const userCredential = await signInWithEmailAndPassword(auth, email, password);
+            alert("ログイン成功！");
+            Cookies.set("user", JSON.stringify(userCredential.user), { expires: 7 });
+            window.location.href = "/top";
+        } catch (error) {
+            console.error("ログイン失敗", error);
+            setError("メールアドレスまたはパスワードが間違っています。");
+        }
+    };
+
+    // Googleでログイン
+    const handleGoogleLogin = async () => {
+        try {
+            const result = await signInWithPopup(auth, provider);
+            alert("Googleログイン成功！");
+            Cookies.set("user", JSON.stringify(result.user), { expires: 7 });
+            window.location.href = "https://pokemon-kaken.vercel.app/sns/post";
+        } catch (error) {
+            console.error("Googleログイン失敗", error);
+            setError("Googleログインに失敗しました。");
+        }
+    };
+
+    // ログアウト
+    const handleLogout = () => {
+        signOut(auth).then(() => {
+            Cookies.remove("user");
+            setUser(null);
+            window.location.href = "/";
+        }).catch((error) => {
+            console.error("ログアウト失敗", error);
+        });
+    };
+
     return (
-        <div
-            style={{
-                backgroundImage: 'url("/images/background.webp")',
-                backgroundSize: "auto", // 画像サイズをそのままに
-                backgroundRepeat: "repeat", // 繰り返して表示
-                backgroundPosition: "top left", // 背景の位置を調整
-                minHeight: "100vh",
-                padding: "0",
-                position: "relative", // 背景画像を親要素に合わせて配置
-            }}
-        >
-
-            {/* 背景画像を全体に適用 */}
-            <div
-                style={{
-                    backgroundImage: 'url("/images/background.png")',
-                    backgroundSize: "cover",
-                    backgroundPosition: "center",
-                    position: "absolute",
-                    top: 0,
-                    left: 0,
-                    width: "100%",
-                    height: "100%",
-                    zIndex: -1, // 背景として表示するため
-                }}
-            ></div>
-
-            {/* ヘッダーを固定表示 */}
+        <div style={{
+            backgroundColor: "#f7f7f7",
+            minHeight: "100vh",
+            display: "flex",
+            flexDirection: "column",
+            justifyContent: "center",
+            alignItems: "center"
+        }}>
+            {/* ヘッダー */}
             <header
                 style={{
-                    backgroundColor: "#FF0000",
+                    backgroundColor: "red",
                     color: "white",
                     textAlign: "center",
                     padding: "20px 0",
                     position: "fixed",
-                    top: 0,
-                    left: 0,
-                    width: "100vw",
-                    zIndex: 1000,
+                    top: "0",
+                    left: "0",
+                    width: "100%",
+                    zIndex: "1000",
                     display: "flex",
                     alignItems: "center",
                     justifyContent: "center",
@@ -58,178 +124,112 @@ const TopPage = () => {
                     }}
                 >
                     <Link href="/">
-                        <div>
-                            <Image
-                                src="/images/gaming.gif"
-                                width={50}
-                                height={50}
-                                alt="s.png"
-                                style={{
-                                    cursor: "pointer",
-                                }}
-                            />
-                        </div>
+                        <Image
+                            src="/images/gaming.gif"
+                            width={50}
+                            height={50}
+                            alt="ホームに戻る"
+                            style={{cursor: "pointer"}}
+                        />
                     </Link>
                 </div>
-                <h1 className="header-title">ポケモンSNSアプリ</h1>
+                <h1 className="header-title">SNSログイン</h1>
             </header>
 
             {/* メインコンテンツ */}
-            <main
-                style={{
-                    textAlign: "center",
-                    padding: "40px 20px",
-                    paddingTop: "180px",  // ヘッダーの高さ分だけ余白を追加                       paddingTop: "80px",
-                    position: "relative",
-                    zIndex: 1,
-                }}
-            >
-                <p style={{marginBottom: "40px", fontSize: "18px", fontWeight: "bold"}}>
-                    各機能へのリンクを選択してください:
-                </p>
-                <div
-                    style={{
-                        display: "flex",
-                        justifyContent: "center",
-                        alignItems: "center",
-                        flexWrap: "wrap",
-                        gap: "50px",
-                    }}
-                >
-                    {[
-                        {href: "/party-builder", label: "パーティ構築", img: "/images/pikacyuuuuu.png"},
-                        {href: "/speed-comparison", label: "クイズ", img: "/images/bata2.png"},
-                        {href: "/damage-calculator", label: "ダメージ計算", img: "/images/bana.png"},
-                        {href: "/zukan", label: "図鑑", img: "/images/riza-don.png"},
-                        {href: "/sns", label: "SNS", img: "/images/kamex.png"},
-                    ].map(({href, label, img}, index) => (
-                        <div key={index} style={{textAlign: "center"}}>
-                            <Link href={href}>
-                                <div style={{display: "inline-block", position: "relative"}}>
-                                    <div className="pokeball">
-                                        <div className="pokeball__top"/>
-                                        <div className="pokeball__bottom"/>
-                                        <div className="pokeball__button"/>
-                                        <div className="pokeball__content">
-                                            <Image
-                                                src={img}
-                                                alt={label}
-                                                layout="fill"
-                                                objectFit="contain"
-                                                style={{
-                                                    display: "block",
-                                                    position: "absolute",
-                                                }}
-                                            />
-                                        </div>
-                                    </div>
-                                </div>
+            <div style={{
+                maxWidth: "400px",
+                padding: "30px",
+                backgroundColor: "white",
+                borderRadius: "10px",
+                boxShadow: "0 4px 6px rgba(0, 0, 0, 0.1)",
+                textAlign: "center",
+                marginTop: "100px"
+            }}>
+                <h2 style={{marginBottom: "20px", color: "#333"}}>ログイン</h2>
+                {user ? (
+                    <button
+                        onClick={handleLogout}
+                        style={{
+                            padding: "10px",
+                            fontSize: "16px",
+                            backgroundColor: "gray",
+                            color: "white",
+                            border: "none",
+                            borderRadius: "5px",
+                            cursor: "pointer",
+                        }}
+                    >
+                        ログアウト
+                    </button>
+                ) : (
+                    <>
+                        <form onSubmit={handleLogin} style={{display: "flex", flexDirection: "column"}}>
+                            <input
+                                type="email"
+                                placeholder="メールアドレス"
+                                value={email}
+                                onChange={(e) => setEmail(e.target.value)}
+                                style={{
+                                    marginBottom: "15px",
+                                    padding: "10px",
+                                    fontSize: "16px",
+                                    border: "1px solid #ccc",
+                                    borderRadius: "5px",
+                                }}
+                                required
+                            />
+                            <input
+                                type="password"
+                                placeholder="パスワード"
+                                value={password}
+                                onChange={(e) => setPassword(e.target.value)}
+                                style={{
+                                    marginBottom: "15px",
+                                    padding: "10px",
+                                    fontSize: "16px",
+                                    border: "1px solid #ccc",
+                                    borderRadius: "5px",
+                                }}
+                                required
+                            />
+                            <button type="submit" style={{
+                                padding: "10px",
+                                fontSize: "16px",
+                                backgroundColor: "red",
+                                color: "white",
+                                border: "none",
+                                borderRadius: "5px",
+                                cursor: "pointer",
+                            }}>
+                                ログイン
+                            </button>
+                        </form>
+                        <button
+                            onClick={handleGoogleLogin}
+                            style={{
+                                marginTop: "15px",
+                                padding: "10px",
+                                fontSize: "16px",
+                                backgroundColor: "#4285F4",
+                                color: "white",
+                                border: "none",
+                                borderRadius: "5px",
+                                cursor: "pointer",
+                            }}
+                        >
+                            Googleでログイン
+                        </button>
+                        {error && <p style={{color: "red", marginTop: "10px"}}>{error}</p>}
+                        <p style={{marginTop: "20px", color: "#555"}}>
+                            アカウントをお持ちでない方は
+                            <Link href="/sns/register" style={{color: "red", textDecoration: "underline", marginLeft: "5px"}}>
+                                新規登録
                             </Link>
-                            <p className="menu-label">{label}</p>
-                        </div>
-                    ))}
-                </div>
-            </main>
-
-            <style jsx>{`
-                @font-face {
-                    font-family: "PokemonGB";
-                    src: url("/font/pokemon-gb-font/pkmn_s.ttf") format("truetype");
-                    font-weight: normal;
-                    font-style: normal;
-                }
-
-                * {
-                    font-family: "PokemonGB", sans-serif;
-                    margin: 0;
-                    padding: 0;
-                    box-sizing: border-box;
-                }
-
-                html, body {
-                    margin: 0;
-                    padding: 0;
-                    width: 100vw;
-                    height: 100vh;
-                    overflow-x: hidden;
-                }
-
-                .header-title {
-                    font-size: 24px;
-                }
-
-                .menu-label {
-                    font-size: 18px;
-                }
-
-                .pokeball {
-                    width: 120px;
-                    height: 120px;
-                    position: relative;
-                    cursor: pointer;
-                    transition: transform 0.3s ease;
-                }
-
-                .pokeball:hover {
-                    transform: scale(1.1);
-                }
-
-                .pokeball__top {
-                    width: 100%;
-                    height: 50%;
-                    background-color: #ff0000;
-                    border-top-left-radius: 50%;
-                    border-top-right-radius: 50%;
-                    position: absolute;
-                    top: 0;
-                    z-index: 1;
-                    transition: transform 0.3s ease;
-                }
-
-                .pokeball__bottom {
-                    width: 100%;
-                    height: 50%;
-                    background-color: #ffffff;
-                    border-bottom-left-radius: 50%;
-                    border-bottom-right-radius: 50%;
-                    position: absolute;
-                    bottom: 0;
-                    z-index: 1;
-                    transition: transform 0.3s ease;
-                }
-
-                .pokeball__button {
-                    position: absolute;
-                    top: 50%;
-                    left: 50%;
-                    width: 20px;
-                    height: 20px;
-                    background-color: #fff;
-                    border: 3px solid #000;
-                    border-radius: 50%;
-                    transform: translate(-50%, -50%);
-                    z-index: 2;
-                    transition: transform 0.3s ease;
-                }
-
-                .pokeball__content {
-                    width: 100%;
-                    height: 100%;
-                    position: absolute;
-                    top: 0;
-                    left: 0;
-                    z-index: 3;
-                    opacity: 0;
-                    pointer-events: none;
-                    transition: opacity 0.3s ease;
-                }
-
-                .pokeball:hover .pokeball__content {
-                    opacity: 1;
-                }
-            `}</style>
+                        </p>
+                    </>
+                )}
+            </div>
         </div>
     );
-};
-
-export default TopPage;
+}
